man: Fix groff style
Replace lowercase `.br` macro with uppercase `.BR` as
described in `man 7 groff_man_style`. Previously in some places
it was uppercase, and in some - lowercase. Now it's uppercase
everywhere.

Just for uniformity with other macros and for the sake
of following the documentation.

Also remove trailing spaces and add a missing `.RE` tag to
fix indentation.

Signed-off-by: Alexander Amelkin <alexander@amelkin.msk.ru>